train/utils.py
import torch
from PIL import Image
from torchvision.utils import make_grid
import os
import cv2



def to_device(data, device):
    """Move tensor(s) to chosen device"""
    if isinstance(data, (list, tuple)):
        return [to_device(x, device) for x in data]
    return data.to(device, non_blocking=True)

class DeviceDataLoader:
    """Wrap a dataloader to move data to a device"""

    def __init__(self, dl, device):
        self.dl = dl
        self.device = device

    def __iter__(self):
        """Yield a batch of data after moving it to device"""
        for b in self.dl:
            yield to_device(b, self.device)

    def __len__(self):
        """Number of batches"""
        return len(self.dl)

def get_default_device():
    return torch.device("cuda" if torch.cuda.is_available() else "cpu")

def save_images(images, path, **kwargs):
    grid = make_grid(images, **kwargs)
    ndarr = grid.permute(1, 2, 0).to("cpu").numpy()
    im = Image.fromarray(ndarr)
    im.save(path)
    
def get(element: torch.Tensor, t: torch.Tensor):
    """
    Get value at index position "t" in "element" and
        reshape it to have the same dimension as a batch of images.
    """
    ele = element.gather(-1, t)
    return ele.reshape(-1, 1, 1, 1)

def setup_log_directory(config):
    '''Log and Model checkpoint directory Setup'''
    
    if os.path.isdir(config.root_log_dir):
        # Get all folders numbers in the root_log_dir
        folder_numbers = [int(folder.replace("version_", "")) for folder in os.listdir(config.root_log_dir)]
        
        # Find the latest version number present in the log_dir
        last_version_number = max(folder_numbers)

        # New version name
        version_name = f"version_{last_version_number + 1}"

    else:
        version_name = config.log_dir

    # Update the training config default directory 
    log_dir        = os.path.join(config.root_log_dir,        version_name)
    checkpoint_dir = os.path.join(config.root_checkpoint_dir, version_name)
    prevs_dir      = os.path.join(config.root_prevs_dir, version_name)

    # Create new directory for saving new experiment version
    os.makedirs(log_dir,        exist_ok=True)
    os.makedirs(checkpoint_dir, exist_ok=True)
    os.makedirs(prevs_dir, exist_ok=True)

    print(f"Logging at: {log_dir}")
    print(f"Model Checkpoint at: {checkpoint_dir}")
    print(f"Pre-visualization at: {prevs_dir}")
    
    return log_dir, checkpoint_dir,prevs_dir, version_name # add version name as the model output

def frames2vid(images, save_path):

    WIDTH = images[0].shape[1]
    HEIGHT = images[0].shape[0]

    fourcc = cv2.VideoWriter_fourcc(*'mp4v')
    video = cv2.VideoWriter(save_path, fourcc, 25, (WIDTH, HEIGHT))

    # Appending the images to the video one by one
    for image in images:
        video.write(image)


    video.release()
    return 



def create_dir_if_nonexist(path):
    '''
    create the path if the path does not exists
    '''
    os.makedirs(path,        exist_ok=True)
    return


def get_version_num(path_dir:str):
    '''
    get the version number given a folder path 
    '''
    if os.path.isdir(path_dir):
        # Get all folders numbers in the root_log_dir
        version_started_folder_names = [each for each in os.listdir(path_dir) if each.startswith("version_") ]

        folder_numbers = [int(folder.replace("version_", "")) for folder in version_started_folder_names]
        
        if len(folder_numbers) == 0:
            version_num = 0
        else:
            # Find the latest version number present in the log_dir
            version_num = max(folder_numbers) + 1
    else:
        version_num = 0

    return str(version_num)



